Upgraded coinmarketcap to v3 framework
