+++ .changeset/strange-glasses-own.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/coinmarketcap-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
