+++ packages/composites/historical-average/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "@chainlink/ea-test-helpers": "workspace:*",
+    "@chainlink/external-adapter-framework": "0.28.0",
     "@types/jest": "27.5.2",
     "@types/node": "16.11.51",
     "@types/supertest": "2.0.12",
