+++ packages/composites/historical-average/src/config/index.ts
@@ -1,6 +1,7 @@
-import * as CoinMarketCap from '@chainlink/coinmarketcap-adapter'
+import { adapter as CoinMarketCap } from '@chainlink/coinmarketcap-adapter'
 import { Requester, util } from '@chainlink/ea-bootstrap'
-import { AdapterImplementation, Config as DefaultConfig } from '@chainlink/ea-bootstrap'
+import { Config as DefaultConfig } from '@chainlink/ea-bootstrap'
+import { Adapter as AdapterImplementation } from '@chainlink/external-adapter-framework/adapter'
 
 export type SourceRequestOptions = { [source: string]: DefaultConfig }
 export const NAME = 'HISTORICAL_AVERAGE'
@@ -10,16 +11,15 @@ export type Config = DefaultConfig & {
   defaultSource?: string
 }
 
-// TODO: adapter implementation types
-export const adapters: AdapterImplementation[] = [CoinMarketCap as unknown as AdapterImplementation]
+export const adapters: AdapterImplementation[] = [CoinMarketCap as AdapterImplementation]
 
-export type Source = (typeof adapters)[number]['NAME']
+export type Source = (typeof adapters)[number]['name']
 
 export const makeConfig = (prefix = ''): Config => {
   const sources: SourceRequestOptions = {}
 
   for (const a of adapters) {
-    const name = a.NAME
+    const name = a.name
     const url = util.getURL(name.toUpperCase())
     if (url) {
       const defaultConfig = Requester.getDefaultConfig(prefix)
