+++ packages/core/legos/tsconfig.json
@@ -134,9 +134,6 @@
       "path": "../../sources/coinmarketcap"
     },
     {
-      "path": "../../sources/coinmarketcap-test"
-    },
-    {
       "path": "../../sources/coinmetrics"
     },
     {
