--- packages/sources/coinmarketcap-test/CHANGELOG.md
@@ -1,73 +1,0 @@
-# @chainlink/coinmarketcap-test-adapter
-
-## 1.2.3
-
-### Patch Changes
-
-- c9cee149e: Bump fw version
-
-## 1.2.2
-
-### Patch Changes
-
-- 957826f14: Fixed request building for empty param lists
-
-## 1.2.1
-
-### Patch Changes
-
-- c35b8ec32: Empty change
-
-## 1.2.0
-
-### Minor Changes
-
-- 17011bb4b: Refactor crypto, volume and marketcap endpoints into one
-
-## 1.1.4
-
-### Patch Changes
-
-- a19738701: Reduced rate limit tiers to account for multiple credits used per request
-
-## 1.1.3
-
-### Patch Changes
-
-- 2fdaa5aa4: Bump v3 version
-
-## 1.1.2
-
-### Patch Changes
-
-- bc8886c63: Bumped framework version
-
-## 1.1.1
-
-### Patch Changes
-
-- 37c0d8883: Bumped v3 framework version
-
-## 1.1.0
-
-### Minor Changes
-
-- d474a5553: Update v3 framework to latest version
-
-## 1.0.2
-
-### Patch Changes
-
-- 65014014d: Bumped v3 framework version. Upgraded typescript to version 4.9.5
-
-## 1.0.1
-
-### Patch Changes
-
-- e303dd766: Add RETH symbol -> ID override
-
-## 1.0.0
-
-### Major Changes
-
-- dd4f3039c: Initial version of the adapter
