--- packages/sources/coinmarketcap-test/src/endpoint/index.ts
@@ -1,4 +1,0 @@
-export { endpoint as crypto } from './crypto'
-export { endpoint as dominance } from './dominance'
-export { endpoint as globalMarketCap } from './globalMarketCap'
-export { endpoint as historical } from './historical'
