--- packages/sources/coinmarketcap/schemas/env.json
@@ -1,18 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/coinmarketcap-adapter.json",
-  "title": "Chainlink External Adapter for Coinmarketcap",
-  "required": ["API_KEY"],
-  "type": "object",
-  "properties": {
-    "API_KEY": {
-      "type": "string",
-      "description": "An API key that can be obtained from [here](https://coinmarketcap.com/api/)",
-      "required": true
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
