--- packages/sources/coinmarketcap/src/adapter.ts
@@ -1,32 +1,0 @@
-import { Builder } from '@chainlink/ea-bootstrap'
-import {
-  ExecuteWithConfig,
-  ExecuteFactory,
-  Config,
-  AdapterRequest,
-  APIEndpoint,
-} from '@chainlink/ea-bootstrap'
-import { makeConfig } from './config'
-import * as endpoints from './endpoint'
-
-export const execute: ExecuteWithConfig<Config, endpoints.TInputParameters> = async (
-  request,
-  context,
-  config,
-) => {
-  return Builder.buildSelector<Config, endpoints.TInputParameters>(
-    request,
-    context,
-    config,
-    endpoints,
-  )
-}
-
-export const endpointSelector = (
-  request: AdapterRequest,
-): APIEndpoint<Config, endpoints.TInputParameters> =>
-  Builder.selectEndpoint<Config, endpoints.TInputParameters>(request, makeConfig(), endpoints)
-
-export const makeExecute: ExecuteFactory<Config, endpoints.TInputParameters> = (config) => {
-  return async (request, context) => execute(request, context, config || makeConfig())
-}
