--- packages/sources/coinmarketcap/src/config/limits.json
@@ -1,26 +1,0 @@
-{
-  "http": {
-    "free": {
-      "rateLimit1m": 30,
-      "rateLimit1h": 13.698630137,
-      "note": "10k credits/month, 730h in a month, ignoring daily limits since they're soft caps"
-    },
-    "hobbyist": {
-      "rateLimit1m": 30,
-      "rateLimit1h": 54.794520548
-    },
-    "startup": {
-      "rateLimit1m": 30,
-      "rateLimit1h": 164.383561644
-    },
-    "standard": {
-      "rateLimit1m": 60,
-      "rateLimit1h": 684.931506849
-    },
-    "professional": {
-      "rateLimit1m": 90,
-      "rateLimit1h": 4109.589041096
-    }
-  },
-  "ws": {}
-}
