+++ packages/sources/coinmarketcap/src/config/overrides.json
