+++ packages/sources/coinmarketcap/src/config/presetids.json
