--- packages/sources/coinmarketcap/src/config/symbols.json
@@ -1,6 +1,0 @@
-{
-  "coinmarketcap": {
-    "PAX": "USDP",
-    "RENFIL": "FIL"
-  }
-}
