+++ packages/sources/coinmarketcap/src/endpoint/crypto.ts
@@ -1,26 +1,66 @@
-import { Requester, Validator, CacheKey, AdapterInputError } from '@chainlink/ea-bootstrap'
-import { NAME as AdapterName } from '../config'
-import type {
-  ExecuteWithConfig,
-  Config,
-  AxiosResponse,
-  AdapterRequest,
-  InputParameters,
-  AdapterBatchResponse,
-} from '@chainlink/ea-bootstrap'
-import overrides from '../config/symbols.json'
+import {
+  CryptoPriceEndpoint,
+  PriceEndpointInputParameters,
+} from '@chainlink/external-adapter-framework/adapter'
+import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
+import { makeLogger, SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
+import { InputParameters } from '@chainlink/external-adapter-framework/validation'
+import { config } from '../config'
+import overrides from '../config/overrides.json'
+import presetIds from '../config/presetids.json'
 
-export const supportedEndpoints = ['crypto', 'price', 'marketcap', 'volume']
-export const batchablePropertyPath = [{ name: 'base' }, { name: 'convert', limit: 120 }]
+const logger = makeLogger('CryptoCMCEndpoint')
 
-export const endpointResultPaths = {
-  crypto: 'price',
-  price: 'price',
-  marketcap: 'market_cap',
-  volume: 'volume_24h',
+interface CryptoRequestParams {
+  base: string
+  quote: string
+  cid?: string
+  slug?: string
+  resultPath: 'price' | 'volume_24h' | 'market_cap'
 }
 
-export interface ResponseSchema {
+const inputParameters = {
+  base: {
+    aliases: ['from', 'coin', 'sym', 'symbol'],
+    description: 'The symbol of symbols of the currency to query',
+    required: true,
+    type: 'string',
+  },
+  quote: {
+    aliases: ['to', 'market', 'convert'],
+    description: 'The symbol of the currency to convert to',
+    required: true,
+    type: 'string',
+  },
+  cid: {
+    description: 'The CMC coin ID (optional to use in place of base)',
+    required: false,
+    type: 'string',
+  },
+  slug: {
+    description: 'The CMC coin ID (optional to use in place of base)',
+    required: false,
+    type: 'string',
+  },
+  resultPath: {
+    description: 'The path to the result within the asset quote in the provider response',
+    required: false,
+    type: 'string',
+    options: ['price', 'volume', 'marketcap'],
+  },
+} satisfies InputParameters & PriceEndpointInputParameters
+
+interface PriceInfo {
+  price: number
+  volume_24h: number
+  percent_change_1h: number
+  percent_change_24h: number
+  percent_change_7d: number
+  percent_change_30d: number
+  market_cap: number
+}
+
+interface ProviderResponseBody {
   data: {
     [key: string]: {
       id: number
@@ -39,16 +79,7 @@ export interface ResponseSchema {
       tags: string[]
       platform: string
       quote: {
-        [key: string]: {
-          price: number
-          volume_24h: number
-          percent_change_1h: number
-          percent_change_24h: number
-          percent_change_7d: number
-          percent_change_30d: number
-          market_cap: number
-          last_updated: string
-        }
+        [key: string]: PriceInfo
       }
     }
   }
@@ -62,215 +93,159 @@ export interface ResponseSchema {
   cost: number
 }
 
-// Coin IDs fetched from the ID map: https://coinmarketcap.com/api/documentation/v1/#operation/getV1CryptocurrencyMap
-const presetIds: { [symbol: string]: number } = {
-  '1INCH': 8104,
-  AAVE: 7278,
-  BAL: 5728,
-  BAT: 1697,
-  BCH: 1831,
-  BNB: 1839,
-  BNT: 1727,
-  BTC: 1,
-  COMP: 5692,
-  CRO: 3635,
-  CRV: 6538,
-  ETC: 1321,
-  ETH: 1027,
-  FNX: 5712,
-  FTT: 4195,
-  HT: 2502,
-  KCS: 2087,
-  KLAY: 4256,
-  KNC: 1982,
-  LEO: 3957,
-  LINK: 1975,
-  MIM: 162,
-  MKR: 1518,
-  OHM: 9067,
-  OHMV2: 9067,
-  OKB: 3897,
-  RCN: 2096,
-  REN: 2539,
-  SNX: 2586,
-  STETH: 8085,
-  SUSHI: 6758,
-  UMA: 5617,
-  UNI: 7083,
-  WOM: 5328,
-  YFI: 5864,
-  RETH: 15060,
+type CryptoEndpointTypes = {
+  Request: {
+    Params: CryptoRequestParams
+  }
+  Response: SingleNumberResultResponse
+  Settings: typeof config.settings
+  Provider: {
+    RequestBody: never
+    ResponseBody: ProviderResponseBody
+  }
 }
 
-export const description = `https://pro-api.coinmarketcap.com/v1/cryptocurrency/quotes/latest
+const chunkArray = <T>(params: T[], size = 120): T[][] =>
+  params.length > size ? [params.slice(0, size), ...chunkArray(params.slice(size), size)] : [params]
 
-**NOTE: the \`price\` endpoint is temporarily still supported, however, is being deprecated. Please use the \`crypto\` endpoint instead.**`
+const resultPathMap = {
+  price: 'price',
+  crypto: 'price',
+  volume: 'volume_24h',
+  marketcap: 'market_cap',
+} as const
 
-export type TInputParameters = {
-  base: string | string[]
-  convert: string | string[]
-  cid: string
-  slug: string
-}
-export const inputParameters: InputParameters<TInputParameters> = {
-  base: {
-    aliases: ['from', 'coin', 'sym', 'symbol'],
-    description: 'The symbol of the currency to query',
-    required: true,
-  },
-  convert: {
-    aliases: ['quote', 'to', 'market'],
-    description: 'The symbol of the currency to convert to',
-    required: true,
-  },
-  cid: {
-    description: 'The CMC coin ID (optional to use in place of base)',
-    required: false,
-    type: 'string',
-  },
-  slug: {
-    description: 'The CMC coin ID (optional to use in place of base)',
-    required: false,
-    type: 'string',
-  },
-}
+const httpTransport = new HttpTransport<CryptoEndpointTypes>({
+  prepareRequests: (params: CryptoRequestParams[], settings) => {
+    const requests = []
+    const groupedParams = {
+      id: [],
+      slug: [],
+      symbol: [],
+    } as Record<string, CryptoRequestParams[]>
 
-const handleBatchedRequest = (
-  jobRunID: string,
-  request: AdapterRequest,
-  response: AxiosResponse<ResponseSchema>,
-  validator: Validator<TInputParameters>,
-  resultPath: string,
-) => {
-  const payload: AdapterBatchResponse = []
-  for (const base in response.data.data) {
-    const originalBase = validator.overrideReverseLookup(
-      AdapterName,
-      'overrides',
-      response.data.data[base].symbol,
-    )
-    for (const quote in response.data.data[base].quote) {
-      const individualRequest = {
-        ...request,
-        data: {
-          ...request.data,
-          base: originalBase.toUpperCase(),
-          convert: quote.toUpperCase(),
-        },
+    for (const param of params) {
+      if (param.cid) {
+        groupedParams.id.push(param)
+      } else if (param.slug) {
+        groupedParams.slug.push(param)
+      } else if (param.base) {
+        groupedParams.symbol.push(param)
+      } else {
+        logger.error(
+          `Params were not able to be classified into ID, Slug or Symbol: (${JSON.stringify(
+            param,
+          )})`,
+        )
+      }
+    }
+
+    for (const [idType, fullList] of Object.entries(groupedParams)) {
+      if (fullList && fullList.length > 0) {
+        // CMC does not support more than 120 unique quotes
+        const chunkedList = chunkArray(fullList, 120)
+
+        // This could be further optimized in cases with more than 120 entries, to make sure that
+        // chunkes are grouped optimally to avoid sending unnecessary converts
+        for (const list of chunkedList) {
+          requests.push({
+            params: list,
+            request: {
+              baseURL: settings.API_ENDPOINT,
+              url: '/cryptocurrency/quotes/latest',
+              headers: {
+                'X-CMC_PRO_API_KEY': settings.API_KEY,
+              },
+              params: {
+                [idType]: [...new Set(list.map((p) => p.cid || p.slug || p.base))].join(','),
+                convert: [...new Set(list.map((p) => p.quote))].join(','),
+              },
+            },
+          })
+        }
+      }
+    }
+
+    return requests
+  },
+  parseResponse: (params, res) => {
+    logger.debug(`CMC api call cost: ${res.data.cost}`)
+
+    // Use the mapping to generate the responses
+    return params.map((p) => {
+      const data = res.data.data[p.cid || p.slug || p.base]
+      if (!data) {
+        return {
+          params: p,
+          response: {
+            statusCode: 502,
+            errorMessage: `Data was not found in response for request: ${JSON.stringify(p)}`,
+          },
+        }
       }
 
-      const result = Requester.validateResultNumber(response.data, [
-        'data',
-        base,
-        'quote',
-        quote,
-        resultPath,
-      ])
+      const dataForQuote = data.quote[p.quote]
+      if (!dataForQuote) {
+        return {
+          params: p,
+          response: {
+            statusCode: 502,
+            errorMessage: `Data for quote "${
+              p.quote
+            }" was not found in response for request: ${JSON.stringify(p)}`,
+          },
+        }
+      }
 
-      payload.push([
-        CacheKey.getCacheKey(individualRequest, Object.keys(inputParameters)),
+      const valueRequested = dataForQuote[p.resultPath]
+      if (valueRequested == null) {
+        return {
+          params: p,
+          response: {
+            statusCode: 502,
+            errorMessage: `Value for "${
+              p.resultPath
+            }" was not found in the quote request: ${JSON.stringify(p)}`,
+          },
+        }
+      }
 
-        individualRequest,
-        result,
-      ])
-    }
-  }
-
-  const results = payload
-  response.data.cost = Requester.validateResultNumber<ResponseSchema>(response.data, [
-    'status',
-    'credit_count',
-  ])
-  return Requester.success(
-    jobRunID,
-    Requester.withResult(response, undefined, results),
-    true,
-    batchablePropertyPath,
-  )
-}
-
-export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
-  const url = 'cryptocurrency/quotes/latest'
-  const validator = new Validator(request, inputParameters, {}, { overrides })
-
-  const jobRunID = validator.validated.id
-  const symbol = validator.overrideSymbol(AdapterName, validator.validated.data.base)
-  // CMC allows a coin name to be specified instead of a symbol
-  const slug = validator.validated.data.slug
-  // CMC allows a coin ID to be specified instead of a symbol
-  const cid = validator.validated.data.cid || ''
-  const convert = validator.validated.data.convert
-  if (!config.apiKey && Array.isArray(convert))
-    throw new AdapterInputError({
-      jobRunID,
-      statusCode: 400,
-      message: ' Free CMCPro API only supports a single symbol to convert',
+      // We're adding multiple results because the same provider endpoint provides values for several adapter endpoints
+      // Price
+      return {
+        params: p,
+        response: {
+          result: dataForQuote[p.resultPath],
+          data: {
+            result: dataForQuote[p.resultPath],
+          },
+        },
+      }
     })
-  const resultPath = (validator.validated.data.resultPath || '').toString()
+  },
+})
 
-  const params: Record<string, string> = {
-    convert: Array.isArray(convert)
-      ? convert.map((symbol) => symbol.toUpperCase()).join(',')
-      : convert.toUpperCase(),
-  }
-  if (cid) {
-    params.id = cid
-  } else if (slug) {
-    params.slug = slug
-  } else if (Array.isArray(symbol)) {
-    let hasIds = true
-    const idsForSymbols = symbol.map((symbol) => {
-      const idForSymbol = presetIds[symbol]
-      if (!idForSymbol) hasIds = false
-      return idForSymbol
-    })
-    if (hasIds) {
-      params.id = idsForSymbols.join(',')
-    } else {
-      params.symbol = symbol.map((s) => s.toUpperCase()).join(',')
-    }
-  } else {
-    const idForSymbol = presetIds[symbol]
-    if (idForSymbol) {
-      params.id = String(idForSymbol)
-    } else {
-      params.symbol = symbol.toUpperCase()
-    }
-  }
+export const endpoint = new CryptoPriceEndpoint<CryptoEndpointTypes>({
+  name: 'crypto',
+  aliases: ['price', 'marketcap', 'volume'],
+  requestTransforms: [
+    (request) => {
+      const { base, cid } = request.requestContext.data
+      const normalizedBase = base.toUpperCase()
+      const idsMap = presetIds as Record<string, number>
+      if (!cid && idsMap[normalizedBase]) {
+        request.requestContext.data.cid = idsMap[normalizedBase].toString()
+      }
 
-  const options = {
-    ...config.api,
-    url,
-    params,
-  }
-  const response = await Requester.request<ResponseSchema>(options)
-  if (Array.isArray(symbol) || Array.isArray(convert))
-    return handleBatchedRequest(jobRunID, request, response, validator, resultPath)
-
-  // CMC API currently uses ID as key in response, when querying with "slug" param
-  const _keyForSlug = (data: ResponseSchema, slug: string) => {
-    if (!data || !data.data) return
-    // First try to find slug key in response (in case the API starts returning keys correctly)
-    if (Object.keys(data.data).includes(slug)) return slug
-    // Fallback to ID
-    const _iEqual = (s1: string, s2: string) => s1.toUpperCase() === s2.toUpperCase()
-    const o = Object.values(data.data).find((o) => _iEqual(o.slug, slug))
-
-    return o && o.id
-  }
-
-  const key = params.id || _keyForSlug(response.data, params.slug || '') || params.symbol
-  const result = Requester.validateResultNumber(response.data, [
-    'data',
-    key,
-    'quote',
-    convert,
-    resultPath,
-  ])
-  return Requester.success(
-    jobRunID,
-    Requester.withResult(response, result),
-    config.verbose,
-    batchablePropertyPath,
-  )
-}
+      if (!request.requestContext.data.resultPath) {
+        const endpoint =
+          (request.body.data as { endpoint: keyof typeof resultPathMap }).endpoint ||
+          request.requestContext.endpointName
+        request.requestContext.data.resultPath = resultPathMap[endpoint]
+      }
+    },
+  ],
+  transport: httpTransport,
+  inputParameters: inputParameters,
+  overrides: overrides.coinmarketcap,
+})
