+++ packages/sources/coinmarketcap/src/global-utils.ts
