+++ packages/sources/coinmarketcap/test/integration/fixtures.ts
