+++ packages/tsconfig.json
@@ -246,9 +246,6 @@
       "path": "./sources/coinmarketcap"
     },
     {
-      "path": "./sources/coinmarketcap-test"
-    },
-    {
       "path": "./sources/coinmetrics"
     },
     {
